6718 FIX Fixed invalid exception handling when unable to initialize the Event Console
When an exception occurs before the configuration was read during initialization
of the Event Console a subsequent exception could occur during shutdown of the
Event Console process. The exception message looked like this:

local variable 'pid_path' referenced before assignment

+++ cmk/ec/main.py
@@ -3971,6 +3971,7 @@ def main():
                                         pathlib.Path(cmk.paths.default_config_dir),
                                         sys.argv)
 
+    pid_path = None
     try:
         cmk.log.open_log(sys.stderr)
         cmk.log.set_verbosity(settings.options.verbosity)
@@ -4072,7 +4073,7 @@ def main():
         bail_out(logger, traceback.format_exc())
 
     finally:
-        if cmk.store.have_lock(str(pid_path)):
+        if pid_path and cmk.store.have_lock(str(pid_path)):
             try:
                 pid_path.unlink()
             except OSError:
